+++ dataclass_generator/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.2.0] - 2019/12/25
+* enable usage of generic types, eg. `Product<T>`
+* **BREAKING CHANGE**: change generated class name from `_Class` to `_$Class`
+
 ## [0.1.0+2] - 2019/12/22
 * update README - repo badges